+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Changed
 
+- Fix issue when updating roles with empty values (cf. [#147](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/147))
+- Support of entitlements with attribute for dimension import
 - Add command for status bar to copy MCP URL
 
 ### Fixed